requirements.txt
Flask
requests
openai
gunicorn
python-dotenv